change layout in other pages, fix buttons
