+++ docker-compose.yml
@@ -8,11 +8,11 @@ services:
     build:
       context: .
       # possible to use development here instead
-      target: development
+      target: production
     ports:
       # development port is usually 3000
-      - "8070:3000"
+      - "8070:8080"
     # for development, uncomment this to map most recent files to the container
-    volumes:
-      - .:/app
+    # volumes:
+    #   - .:/app
     restart: always
