+++ src/components/Layout/Footer.tsx
@@ -1,40 +1,34 @@
-import { Box } from "@mui/material";
-import React from 'react';
-const Footer: React.FC = () => {
+import { Box, IconButton, Link, Typography } from '@mui/material';
+import React, { useState } from 'react';
+import logoArasaac from './logo_ARASAAC.png';
+import { Trans, useTranslation } from 'react-i18next';
+import Clipboard from "../../utils/Clipboard";
+import ClipboardIcon from '@mui/icons-material/ContentPasteGo';
 
-        return (
-            <Box sx={{
-                background: '#ffffff',
-                margin: '80px auto 30px auto',
-                textAlign: 'center',
-            }}>
-                <Box component="ul" sx={{
-                    padding: 0,
-                    textAlign: 'center',
-                    margin: '5px',
-                    color: '#999999',
-                    '& li': { display: 'inline-block', padding: '5px 10px' },
-                    '& a': {
-                        color: 'inherit',
-                        fontSize: '12px',
-                        textDecoration: 'none',
-                        '&:hover': {
-                            color: '#555555',
-                        }
-                    }
-                }}>
-                    <li>
-                        <a href="https://kits.blog/impressum/" target="_blank" rel="noopener noreferrer">Impressum</a>
-                    </li>
-                    <li>
-                        <a href="https://kits.blog/datenschutzerklaerung/" target="_blank" rel="noopener noreferrer">Datenschutz</a>
-                    </li>
-                    <li>
-                        <a href="https://github.com/kitsteam/wp-pictogram-viewer" target="_blank" rel="noopener noreferrer">PictoSearch@Github</a>
-                    </li>
-                </Box>
-            </Box>
-        );
+const Footer: React.FC = () => {
+    const { t } = useTranslation();
+    const [copiedLicenseToClipboard, setCopiedLicenseToClipboard] = useState(false);
+    const onCopyLicenseToClipboard = () => {
+      Clipboard.copyText(t('search.clipboardLicense'));
+      setCopiedLicenseToClipboard(true);
+      setTimeout(() => setCopiedLicenseToClipboard(false), 2000);
+    }
+
+    return (
+        <Box className="app-footer">
+            <Link href="http://www.arasaac.org">
+              <img src={logoArasaac} alt="Logo ARASSAC" />
+            </Link>
+            <Typography variant="body2" sx={{ opacity: 0.6 }}>
+              {Clipboard.hasSupport() && <IconButton sx={{ marginLeft: -1 }} size="small" onClick={onCopyLicenseToClipboard} color={copiedLicenseToClipboard ? 'success' : 'default'}><ClipboardIcon /></IconButton>}
+
+              <Trans i18nKey="search.license">
+                Sergio Palao (Urheber), ARASAAC (<Link href="http://www.arasaac.org">arasaac.org</Link>),
+                Regierung von Aragón in Spanien (Eigentümer), <Link href="https://creativecommons.org/licenses/by-nc-sa/4.0/deed.en">CC BY-SA-NC 4.0</Link>.
+              </Trans>
+            </Typography>
+        </Box>
+    );
 }
